Move backend remaining tests
