+++ packages/backend/package.json
@@ -2,8 +2,8 @@
   "name": "@explorer/backend",
   "version": "0.1.0",
   "private": true,
-  "main": "build/src/index.js",
-  "types": "build/src/index.d.ts",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
   "engines": {
     "node": "^16.0.0",
     "yarn": ">=1.22.0"
@@ -11,13 +11,15 @@
   "scripts": {
     "format": "prettier --check .",
     "format:fix": "prettier --write .",
-    "lint": "eslint --ext .ts --max-warnings 0 src test",
+    "lint": "eslint --ext .ts --max-warnings 0 src",
     "lint:fix": "yarn lint --fix",
     "typecheck": "tsc --noEmit",
+    "prebuild": "yarn clean",
     "build": "tsc",
+    "clean": "rm -rf build",
     "test": "mocha",
     "test:watch": "mocha --watch --watch-files src/**/*.ts,test/**/*.ts",
-    "start": "node -r source-map-support/register build/src",
+    "start": "node -r source-map-support/register build",
     "dev": "node -r esbuild-register src"
   },
   "dependencies": {
