+++ packages/backend/src/core/PerpetualRollupSyncService.test.ts
@@ -1,20 +1,20 @@
 import { EthereumAddress, Hash256 } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { FinalizeExitEventsCollector } from '../../src/core/collectors/FinalizeExitEventsCollector'
-import { ForcedEventsCollector } from '../../src/core/collectors/ForcedEventsCollector'
-import type { PageCollector } from '../../src/core/collectors/PageCollector'
-import type { PageMappingCollector } from '../../src/core/collectors/PageMappingCollector'
-import { PerpetualRollupStateTransitionCollector } from '../../src/core/collectors/PerpetualRollupStateTransitionCollector'
-import { UserRegistrationCollector } from '../../src/core/collectors/UserRegistrationCollector'
-import type { VerifierCollector } from '../../src/core/collectors/VerifierCollector'
-import { PerpetualRollupSyncService } from '../../src/core/PerpetualRollupSyncService'
-import { PerpetualRollupUpdater } from '../../src/core/PerpetualRollupUpdater'
-import { BlockRange } from '../../src/model'
-import { StateTransitionRecord } from '../../src/peripherals/database/StateTransitionRepository'
-import { decodedFakePages, fakePages } from '../../src/test/fakes'
-import { mock } from '../../src/test/mock'
-import { Logger } from '../../src/tools/Logger'
+import { BlockRange } from '../model'
+import { StateTransitionRecord } from '../peripherals/database/StateTransitionRepository'
+import { decodedFakePages, fakePages } from '../test/fakes'
+import { mock } from '../test/mock'
+import { Logger } from '../tools/Logger'
+import { FinalizeExitEventsCollector } from './collectors/FinalizeExitEventsCollector'
+import { ForcedEventsCollector } from './collectors/ForcedEventsCollector'
+import type { PageCollector } from './collectors/PageCollector'
+import type { PageMappingCollector } from './collectors/PageMappingCollector'
+import { PerpetualRollupStateTransitionCollector } from './collectors/PerpetualRollupStateTransitionCollector'
+import { UserRegistrationCollector } from './collectors/UserRegistrationCollector'
+import type { VerifierCollector } from './collectors/VerifierCollector'
+import { PerpetualRollupSyncService } from './PerpetualRollupSyncService'
+import { PerpetualRollupUpdater } from './PerpetualRollupUpdater'
 
 const noop = async () => {}
