+++ packages/backend/src/core/PerpetualRollupUpdater.test.ts
@@ -18,16 +18,16 @@ import {
 } from '@explorer/types'
 import { expect, mockFn } from 'earljs'
 
-import { PerpetualRollupUpdater } from '../../src/core/PerpetualRollupUpdater'
-import { EMPTY_STATE_HASH } from '../../src/core/PerpetualValidiumUpdater'
-import { ForcedTransactionRepository } from '../../src/peripherals/database/ForcedTransactionRepository'
-import type { MerkleTreeRepository } from '../../src/peripherals/database/MerkleTreeRepository'
-import { PageRepository } from '../../src/peripherals/database/PageRepository'
-import { StateTransitionRecord } from '../../src/peripherals/database/StateTransitionRepository'
-import { StateUpdateRepository } from '../../src/peripherals/database/StateUpdateRepository'
-import type { EthereumClient } from '../../src/peripherals/ethereum/EthereumClient'
-import { mock } from '../../src/test/mock'
-import { Logger } from '../../src/tools/Logger'
+import { ForcedTransactionRepository } from '../peripherals/database/ForcedTransactionRepository'
+import type { MerkleTreeRepository } from '../peripherals/database/MerkleTreeRepository'
+import { PageRepository } from '../peripherals/database/PageRepository'
+import { StateTransitionRecord } from '../peripherals/database/StateTransitionRepository'
+import { StateUpdateRepository } from '../peripherals/database/StateUpdateRepository'
+import type { EthereumClient } from '../peripherals/ethereum/EthereumClient'
+import { mock } from '../test/mock'
+import { Logger } from '../tools/Logger'
+import { PerpetualRollupUpdater } from './PerpetualRollupUpdater'
+import { EMPTY_STATE_HASH } from './PerpetualValidiumUpdater'
 
 const emptyState: State = {
   positionRoot: PedersenHash.ZERO,
