+++ packages/backend/src/core/PerpetualValidiumUpdater.test.ts
@@ -18,17 +18,17 @@ import {
 } from '@explorer/types'
 import { expect, mockFn } from 'earljs'
 
+import { ForcedTransactionRepository } from '../peripherals/database/ForcedTransactionRepository'
+import type { MerkleTreeRepository } from '../peripherals/database/MerkleTreeRepository'
+import { StateTransitionRecord } from '../peripherals/database/StateTransitionRepository'
+import { StateUpdateRepository } from '../peripherals/database/StateUpdateRepository'
+import { EthereumClient } from '../peripherals/ethereum/EthereumClient'
+import { mock } from '../test/mock'
+import { Logger } from '../tools/Logger'
 import {
   EMPTY_STATE_HASH,
   PerpetualValidiumUpdater,
-} from '../../src/core/PerpetualValidiumUpdater'
-import { ForcedTransactionRepository } from '../../src/peripherals/database/ForcedTransactionRepository'
-import type { MerkleTreeRepository } from '../../src/peripherals/database/MerkleTreeRepository'
-import { StateTransitionRecord } from '../../src/peripherals/database/StateTransitionRepository'
-import { StateUpdateRepository } from '../../src/peripherals/database/StateUpdateRepository'
-import { EthereumClient } from '../../src/peripherals/ethereum/EthereumClient'
-import { mock } from '../../src/test/mock'
-import { Logger } from '../../src/tools/Logger'
+} from './PerpetualValidiumUpdater'
 
 const emptyState: State = {
   positionRoot: PedersenHash.ZERO,
