+++ packages/backend/src/core/StateUpdater.test.ts
@@ -2,14 +2,14 @@ import { MerkleTree, PositionLeaf } from '@explorer/state'
 import { AssetId, Hash256, PedersenHash, StarkKey } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { StateUpdater } from '../../src/core/StateUpdater'
-import { ForcedTransactionRepository } from '../../src/peripherals/database/ForcedTransactionRepository'
-import type { MerkleTreeRepository } from '../../src/peripherals/database/MerkleTreeRepository'
-import { StateTransitionRecord } from '../../src/peripherals/database/StateTransitionRepository'
-import { StateUpdateRepository } from '../../src/peripherals/database/StateUpdateRepository'
-import type { EthereumClient } from '../../src/peripherals/ethereum/EthereumClient'
-import { mock } from '../../src/test/mock'
-import { Logger } from '../../src/tools/Logger'
+import { ForcedTransactionRepository } from '../peripherals/database/ForcedTransactionRepository'
+import type { MerkleTreeRepository } from '../peripherals/database/MerkleTreeRepository'
+import { StateTransitionRecord } from '../peripherals/database/StateTransitionRepository'
+import { StateUpdateRepository } from '../peripherals/database/StateUpdateRepository'
+import type { EthereumClient } from '../peripherals/ethereum/EthereumClient'
+import { mock } from '../test/mock'
+import { Logger } from '../tools/Logger'
+import { StateUpdater } from './StateUpdater'
 
 const EMPTY_STATE_HASH = PedersenHash(
   '52ddcbdd431a044cf838a71d194248640210b316d7b1a568997ecad9dec9626'
