+++ packages/backend/src/core/collectors/FinalizeExitEventsCollector.test.ts
@@ -1,12 +1,10 @@
 import { EthereumAddress, Hash256, Timestamp } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { LogWithdrawalPerformed } from '../../../src/core/collectors/events'
-import { FinalizeExitEventsCollector } from '../../../src/core/collectors/FinalizeExitEventsCollector'
-import { BlockRange } from '../../../src/model'
-import { ForcedTransactionRepository } from '../../../src/peripherals/database/ForcedTransactionRepository'
-import { TransactionStatusRepository } from '../../../src/peripherals/database/TransactionStatusRepository'
-import { EthereumClient } from '../../../src/peripherals/ethereum/EthereumClient'
+import { BlockRange } from '../../model'
+import { ForcedTransactionRepository } from '../../peripherals/database/ForcedTransactionRepository'
+import { TransactionStatusRepository } from '../../peripherals/database/TransactionStatusRepository'
+import { EthereumClient } from '../../peripherals/ethereum/EthereumClient'
 import {
   fakeBlock,
   fakeExit,
@@ -14,8 +12,10 @@ import {
   fakeForcedUpdatesVerified,
   fakeInt,
   fakeTimestamp,
-} from '../../../src/test/fakes'
-import { mock } from '../../../src/test/mock'
+} from '../../test/fakes'
+import { mock } from '../../test/mock'
+import { LogWithdrawalPerformed } from './events'
+import { FinalizeExitEventsCollector } from './FinalizeExitEventsCollector'
 
 const blockRange = new BlockRange([
   {
